COMDEVNL-1157 -- Add EPS and Giropay to 2.2-and-lower module and upgrade it to API v2

Approved-by: mark smit <m.smit@hostingondemand.nl>

+++ admin/controller/payment/mollie/base.php
@@ -42,13 +42,13 @@
  * @property URL                          $url
  * @property User                         $user
  */
+
 use Mollie\Api\Exceptions\ApiException;
 use Mollie\Api\Exceptions\IncompatiblePlatform;
 use Mollie\Api\MollieApiClient;
+require_once(dirname(DIR_SYSTEM) . "/catalog/controller/payment/mollie/helper.php");
 
-require_once(dirname(DIR_SYSTEM) . "/catalog/controller/extension/payment/mollie/helper.php");
-
-class ControllerExtensionPaymentMollieBase extends Controller
+class ControllerPaymentMollieBase extends Controller
 {
 	// Current module name - should be overwritten by subclass using one of the MollieHelper::MODULE_NAME_* values.
 	const MODULE_NAME = NULL;
@@ -61,7 +61,7 @@ class ControllerExtensionPaymentMollieBase extends Controller
 
 	/**
 	 * @param int $store The Store ID
-	 * @return MollieApiClient
+	 * @return MollieAPIClient
 	 */
 	protected function getAPIClient ($store = 0)
 	{
@@ -102,11 +102,13 @@ class ControllerExtensionPaymentMollieBase extends Controller
 	 */
 	public function cleanUp()
 	{
-		$adminThemeDir = DIR_APPLICATION . 'view/template/extension/payment/';
-		$catalogThemeDir = DIR_CATALOG . 'view/theme/default/template/extension/payment/';
+		$adminThemeDir = DIR_APPLICATION . 'view/template/payment/';
+		$catalogThemeDir = DIR_CATALOG . 'view/theme/default/template/payment/';
 
 		// Remove old template from previous version.
-		unlink($adminThemeDir . 'mollie_2.tpl');
+		if(file_exists($adminThemeDir . 'mollie_2.tpl')) {
+			unlink($adminThemeDir . 'mollie_2.tpl');
+		}
 
 		if (MollieHelper::isOpenCart3x()) {
 			unlink($adminThemeDir . 'mollie_1.tpl');
@@ -141,8 +143,8 @@ class ControllerExtensionPaymentMollieBase extends Controller
 			$extensions->install("payment", "mollie_" . $module_name);
 
 			// Set permissions.
-			$this->model_user_user_group->addPermission($user_id, "access", "extension/payment/mollie_" . $module_name);
-			$this->model_user_user_group->addPermission($user_id, "modify", "extension/payment/mollie_" . $module_name);
+			$this->model_user_user_group->addPermission($user_id, "access", "payment/mollie_" . $module_name);
+			$this->model_user_user_group->addPermission($user_id, "modify", "payment/mollie_" . $module_name);
 		}
 	}
 
@@ -175,20 +177,20 @@ class ControllerExtensionPaymentMollieBase extends Controller
 	 */
 	public function index ()
 	{
-		// Double-check if clean-up has been done - For upgrades
-		$adminThemeDir = DIR_APPLICATION . 'view/template/extension/payment/';
+		$adminThemeDir = DIR_APPLICATION . 'view/template/payment/';
 		if (MollieHelper::isOpenCart3x() || MollieHelper::isOpenCart2x()) {
 			if(file_exists($adminThemeDir . 'mollie_1.tpl')) {
 				$this->cleanUp();
 			}
-		} else {
+		}
+		else {
 			if(file_exists($adminThemeDir . 'mollie.tpl')) {
 				$this->cleanUp();
 			}
 		}
 		
 		// Load essential models
-		$this->load->language("extension/payment/mollie");
+		$this->load->language("payment/mollie");
 		$this->load->model("setting/setting");
 		$this->load->model("setting/store");
 		$this->load->model("localisation/order_status");
@@ -233,7 +235,7 @@ class ControllerExtensionPaymentMollieBase extends Controller
 		}
 
 		// Set data for template
-		$data['api_check_url']          = $this->url->link("extension/payment/mollie_" . static::MODULE_NAME . '/validate_api_key', $this->getTokenUriPart(), "SSL");
+		$data['api_check_url']          = $this->url->link("payment/mollie_" . static::MODULE_NAME . '/validate_api_key', $this->getTokenUriPart(), "SSL");
 		$data['heading_title']          = $this->language->get("heading_title");
 		$data['title_global_options']   = $this->language->get("title_global_options");
 		$data['title_payment_status']   = $this->language->get("title_payment_status");
@@ -351,13 +353,13 @@ class ControllerExtensionPaymentMollieBase extends Controller
 		);
 
 		$data['breadcrumbs'][] = array(
-			"href"      => $this->url->link("extension/payment/mollie_" . static::MODULE_NAME, $this->getTokenUriPart(), "SSL"),
+			"href"      => $this->url->link("payment/mollie_" . static::MODULE_NAME, $this->getTokenUriPart(), "SSL"),
 			"text"      => $this->language->get("heading_title"),
 			"separator" => " :: ",
 		);
 
 		// Form action url
-		$data['action'] = $this->url->link("extension/payment/mollie_" . static::MODULE_NAME, $this->getTokenUriPart(), "SSL");
+		$data['action'] = $this->url->link("payment/mollie_" . static::MODULE_NAME, $this->getTokenUriPart(), "SSL");
 		$data['cancel'] = $this->url->link($this->getExtensionsUri(), "type=payment&" . $this->getTokenUriPart(), "SSL");
 
 		// Load global settings. Some are prefixed with mollie_ideal_ for legacy reasons.
@@ -464,9 +466,9 @@ class ControllerExtensionPaymentMollieBase extends Controller
 			
 		}
 
-		$template = 'extension/payment/mollie';
+		$template = 'payment/mollie';
 		if (!MollieHelper::isOpenCart2x()) {
-			$template = 'extension/payment/mollie_1';
+			$template = 'payment/mollie_1';
 		}
 
 		$this->renderTemplate($template, $data, array(
@@ -476,10 +478,7 @@ class ControllerExtensionPaymentMollieBase extends Controller
 		));
 	}
 
-    /**
-     *
-     */
-    public function validate_api_key()
+	public function validate_api_key()
 	{
 		$json = array(
 			'error' => false,
@@ -534,7 +533,7 @@ class ControllerExtensionPaymentMollieBase extends Controller
 	 */
 	private function validate ($store = 0)
 	{
-		if (!$this->user->hasPermission("modify", "extension/payment/mollie_" . static::MODULE_NAME))
+		if (!$this->user->hasPermission("modify", "payment/mollie_" . static::MODULE_NAME))
 		{
 			$this->error['warning'] = $this->language->get("error_permission");
 		}
@@ -572,9 +571,9 @@ class ControllerExtensionPaymentMollieBase extends Controller
 			$client->methods->all();
 
 			return '<span style="color: green">OK</span>';
-		} catch (Mollie\Api\Exceptions\ApiException_IncompatiblePlatform $e) {
+		} catch (IncompatiblePlatform $e) {
 			return '<span style="color:red">' . $e->getMessage() . ' You can ask your hosting provider to help with this.</span>';
-		} catch (Mollie\Api\Exceptions\ApiException $e) {
+		} catch (ApiException $e) {
 			return '<span style="color:red">'
 				. '<strong>Communicating with Mollie failed:</strong><br/>'
 				. htmlspecialchars($e->getMessage())
